+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Our Keycloak version is working well with PostgreSQL database. For using other S
 
 ### Fixed
 - Correct problems related with autoupdated IdP
+- 
+### Changed
+- Changes related to View Groups from account console PR [PR-7933](https://github.com/keycloak/keycloak/pull/7933)
 
 ## [16.1.0-2.11] - 2022-09-01